What: split client server for go1.8
Why:

  * httptest.Client is undefined <go1.8

How:

  * nop
